moved lib test to lib, and added docker test
