+++ Dockerfile.test
@@ -1,0 +1,6 @@
+FROM node:13.14-alpine
+
+COPY entrypoint.sh /entrypoint.sh
+
+# Executes `entrypoint.sh` when the Docker container starts up
+ENTRYPOINT ["/entrypoint.sh"]
