+++ entrypoint.sh
@@ -1,0 +1,2 @@
+ls -la
+pwd
