+++ lib/smartvault_lib.js
@@ -28,7 +28,7 @@ class SmartVault {
     constructor(config, dir) {
         console.log("Connecting to ", config.RPC_URL, dir);
         this.config = config;
-        this.localDir = dir;
+        this.localDir = dir || ".";
         this.harmonyClient = new HarmonyClient(config.RPC_URL, config.ENS_ADDRESS, config.EXPLORER_URL);    
         this.relayClient = new RelayerFactory(config.API_URL, config.ENV);    
         this.walletData = {created: false};
