+++ lib/test/harmony_client_test.js
@@ -1,5 +1,5 @@
 
-const HarmonyClient = require("../../lib/harmony_client");
+const HarmonyClient = require("../harmony_client");
 const chai = require('chai')
 const expect = chai.expect
 chai.use(require('chai-as-promised'))
@@ -9,7 +9,7 @@ describe("Library test", (accounts) => {
   it("checks for name available", async () => {
     var client = new HarmonyClient("https://api.s3.b.hmny.io", "0x4fb1C434101ced0773a3bc77D541B3465023639f");
     var available = await client.isNameAvailable("hashmesan000.crazy.one", 31536000);
-    console.log("Available?", available);
+    //console.log("Available?", available);
     return true;
   });
