+++ lib/test/smartvault_lib_test.js
@@ -5,12 +5,12 @@ chai.use(require('chai-as-promised'))
 var assert = require('assert');
 
 const config = require("../../webclient/src/config");
-const SmartVault = require("../../lib/smartvault_lib");
-const HarmonyClient = require("../../lib/harmony_client");
+const SmartVault = require("../smartvault_lib");
+const HarmonyClient = require("../harmony_client");
 const Web3 = require("web3");
 const sleep = (sec) => new Promise(resolve => setTimeout(resolve, 1000 * sec))
-const wallet = require("../../lib/wallet");
-const RelayerClient = require('../../lib/relayer_client');
+const wallet = require("../wallet");
+const RelayerClient = require('../relayer_client');
 const Web3EthAccounts = require('web3-eth-accounts');
 
 describe("SmartVault test", () => {
