+++ lib/test/wallet_generation_test.js
@@ -1,5 +1,5 @@
 
-const wallet = require("../../lib/wallet");
+const wallet = require("../wallet");
 const base64 = require("@ethersproject/base64");
 var assert = require('assert');
 const web3utils = require("web3-utils");
